Address lint review findings
Split oversized benchmark and pipeline modules to remove broad
too-many-lines suppressions, narrow protocol stubs, and make the lint
target declare uv as a direct prerequisite.
