+++ benchmarks/pipeline_throughput_runner.py
@@ -1,0 +1,174 @@
+"""Command construction and execution for pipeline throughput benchmarks."""
+
+from __future__ import annotations
+
+import dataclasses as dc
+import json
+import shlex
+import shutil
+import subprocess  # noqa: S404  # benchmark runner intentionally invokes external tooling
+import typing as typ
+
+from benchmarks._benchmark_types import (
+    HyperfineConfig,
+    PipelineBenchmarkConfig,
+    PipelineBenchmarkRunResult,
+    PipelineBenchmarkScenario,
+)
+
+if typ.TYPE_CHECKING:
+    import collections.abc as cabc
+    import pathlib as pth
+
+
+def render_prefixed_command(
+    *,
+    command: cabc.Sequence[str],
+    env: cabc.Mapping[str, str],
+) -> str:
+    """Render a shell command with deterministic environment prefixes."""
+    env_tokens = [f"{key}={shlex.quote(value)}" for key, value in sorted(env.items())]
+    command_text = shlex.join(list(command))
+    if not env_tokens:
+        return command_text
+    return f"{' '.join(env_tokens)} {command_text}"
+
+
+def _build_worker_command(
+    *,
+    scenario: PipelineBenchmarkScenario,
+    worker_path: pth.Path,
+    uv_bin: str,
+) -> list[str]:
+    """Build the worker invocation for one benchmark scenario."""
+    command = [
+        uv_bin,
+        "run",
+        "python",
+        str(worker_path),
+        "--payload-bytes",
+        str(scenario.payload_bytes),
+        "--stages",
+        str(scenario.stages),
+    ]
+    if scenario.with_line_callbacks:
+        command.append("--line-callbacks")
+    return command
+
+
+def _resolve_executable(name: str) -> str:
+    """Resolve an executable to an absolute path."""
+    resolved = shutil.which(name)
+    if resolved is None:
+        msg = f"required executable is not on PATH: {name}"
+        raise FileNotFoundError(msg)
+    return resolved
+
+
+def build_hyperfine_command(*, config: PipelineBenchmarkConfig) -> list[str]:
+    """Construct a hyperfine command vector for configured scenarios."""
+    hyperfine_config = HyperfineConfig(
+        warmup=config.warmup,
+        runs=config.runs,
+        hyperfine_bin=config.hyperfine_bin,
+    )
+    if not config.scenarios:
+        msg = "at least one benchmark scenario is required"
+        raise ValueError(msg)
+
+    command = [
+        hyperfine_config.hyperfine_bin,
+        "--export-json",
+        str(config.output_path),
+        "--warmup",
+        str(hyperfine_config.warmup),
+        "--runs",
+        str(hyperfine_config.runs),
+    ]
+    for scenario in config.scenarios:
+        worker_command = _build_worker_command(
+            scenario=scenario,
+            worker_path=config.worker_path,
+            uv_bin=config.uv_bin,
+        )
+        command.append(
+            render_prefixed_command(
+                command=worker_command,
+                env={"CUPRUM_STREAM_BACKEND": scenario.backend},
+            ),
+        )
+    return command
+
+
+def _write_dry_run_payload(
+    *,
+    output_path: pth.Path,
+    command: cabc.Sequence[str],
+    scenarios: cabc.Sequence[PipelineBenchmarkScenario],
+    rust_available: bool,
+) -> None:
+    """Write dry-run benchmark metadata to JSON."""
+    output_path.parent.mkdir(parents=True, exist_ok=True)
+    payload = {
+        "dry_run": True,
+        "rust_available": rust_available,
+        "command": list(command),
+        "scenarios": [scenario.as_dict() for scenario in scenarios],
+    }
+    output_path.write_text(
+        json.dumps(payload, indent=2, sort_keys=True),
+        encoding="utf-8",
+    )
+
+
+def _execute_hyperfine_benchmark(
+    command: list[str],
+    output_path: pth.Path,
+) -> None:
+    """Execute the hyperfine benchmark command."""
+    command[0] = _resolve_executable(command[0])
+    output_path.parent.mkdir(parents=True, exist_ok=True)
+    subprocess.run(  # noqa: S603  # command built from fixed executable + controlled args
+        command,
+        check=True,
+        timeout=1800,
+    )
+
+
+def _prepare_benchmark_command_config(
+    *,
+    config: PipelineBenchmarkConfig,
+) -> PipelineBenchmarkConfig:
+    """Prepare command rendering config, resolving the `uv` launcher when needed."""
+    if config.dry_run:
+        return config
+    return dc.replace(config, uv_bin=_resolve_executable(config.uv_bin))
+
+
+def run_pipeline_benchmarks(
+    *, config: PipelineBenchmarkConfig
+) -> PipelineBenchmarkRunResult:
+    """Execute pipeline benchmarks or write a dry-run benchmark plan."""
+    command_config = _prepare_benchmark_command_config(config=config)
+    command = build_hyperfine_command(config=command_config)
+
+    if config.dry_run:
+        _write_dry_run_payload(
+            output_path=config.output_path,
+            command=command,
+            scenarios=config.scenarios,
+            rust_available=config.rust_available,
+        )
+    else:
+        _execute_hyperfine_benchmark(
+            command=command,
+            output_path=config.output_path,
+        )
+
+    return PipelineBenchmarkRunResult(
+        dry_run=config.dry_run,
+        command=tuple(command),
+        output_path=config.output_path,
+        rust_available=config.rust_available,
+        scenarios=config.scenarios,
+    )
