+++ cuprum/adapters/tracing_adapter.py
@@ -92,7 +92,8 @@ class Span(typ.Protocol):
             Attribute value (string, int, float, bool, or list thereof).
 
         """
-        raise NotImplementedError
+        # PEP 544 protocol stub.
+        ...  # pylint: disable=unnecessary-ellipsis
 
     def add_event(
         self,
@@ -109,7 +110,8 @@ class Span(typ.Protocol):
             Optional attributes for the event.
 
         """
-        raise NotImplementedError
+        # PEP 544 protocol stub.
+        ...  # pylint: disable=unnecessary-ellipsis
 
     def set_status(self, *, ok: bool) -> None:
         """Set the span status.
@@ -120,11 +122,13 @@ class Span(typ.Protocol):
             True if the operation succeeded, False otherwise.
 
         """
-        raise NotImplementedError
+        # PEP 544 protocol stub.
+        ...  # pylint: disable=unnecessary-ellipsis
 
     def end(self) -> None:
         """End the span, recording its duration."""
-        raise NotImplementedError
+        # PEP 544 protocol stub.
+        ...  # pylint: disable=unnecessary-ellipsis
 
 
 class Tracer(typ.Protocol):
@@ -154,7 +158,8 @@ class Tracer(typ.Protocol):
             A span that must be ended by calling :meth:`Span.end`.
 
         """
-        raise NotImplementedError
+        # PEP 544 protocol stub.
+        ...  # pylint: disable=unnecessary-ellipsis
 
 
 @dc.dataclass(slots=True)
